# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Courses API",
        "description": "API REST da plataforma de cursos online AulasViolão. Autenticação via JWT Bearer Token.",
        "contact": {
            "name": "Big Space Creative",
            "email": "admin@digitalcourses.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Exibir aula",
                "description": "Retorna detalhes de uma aula. Aulas pagas exigem autenticação com plano Premium (ou role admin/instructor).",
                "operationId": "lessonShow",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID da aula",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aula encontrada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — plano Premium necessário"
                    },
                    "404": {
                        "description": "Aula não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Atualizar aula",
                "description": "Atualiza dados de uma aula existente. Requer autenticação (instructor ou admin).",
                "operationId": "lessonUpdate",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID da aula",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 5000
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 2048
                                    },
                                    "duration_in_minutes": {
                                        "type": "integer",
                                        "nullable": true,
                                        "maximum": 1440,
                                        "minimum": 1
                                    },
                                    "is_free_preview": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aula atualizada com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão (apenas admin ou instructor)"
                    },
                    "404": {
                        "description": "Aula não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Excluir aula",
                "description": "Remove permanentemente uma aula. Requer autenticação (admin).",
                "operationId": "lessonDestroy",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID da aula",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula excluída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aula excluída com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Aula não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/materials/{material}": {
            "delete": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Excluir material",
                "description": "Remove permanentemente um material.",
                "operationId": "materialDestroy",
                "parameters": [
                    {
                        "name": "material",
                        "in": "path",
                        "description": "ID do material",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material excluído"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Material não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT obtido no login: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}